backend/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# AI & ML Services
anthropic==0.49.0
requests==2.31.0

# Elasticsearch
elasticsearch==8.12.0

# Environment & Config
python-dotenv==1.0.0

# Data / ML
numpy==1.26.3
scikit-learn>=1.4.0
